requirements.txt
fastapi==0.115.0
starlette==0.38.2
pydantic==2.9.0
PyJWT==2.8.0
cryptography==42.0.7
uvicorn==0.30.1
aiosmtplib==3.0.1
bcrypt==4.0.1
email-validator==2.1.1
